Particle Fire
Using the SDL2 Visual api to render buffered pixels as RGBA values that
randomize and fluctuate while changing colors.

Creation of memory buffers to house said pixels along with SDL

SDL window compliments a game loop which re-renders and updates the
screen more than 25fps

The event loop waits for an action such as a mouse click of the close
button to destroy the SDL window object.

Finally the SDL window  comes to a close when the  window is closed.